requirements.txt
Django==4.2
django-environ==0.4.4
django-widget-tweaks==1.4.1
gunicorn==20.1.0
psycopg2-binary==2.9.7
django-debug-toolbar==4.3

